requirements.txt
Django==4.2.7
django-simple-history==3.4.0
Pillow